services/ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.7.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime==1.20.1
insightface==0.7.3
Pillow==10.4.0
httpx==0.27.2
SQLAlchemy==2.0.46
alembic==1.14.0
mediapipe==0.10.20
